requirements.txt
pika>=0.10.0
python-dateutil>=2.5.3
pytest-cov>=2.2.1
pytest>=2.9.1